Introduce JSON outputting for Merlin callgraph
**Motivation:**

We need to output the Merlin callgraph in a format that is comparable to
other callgraphs from other papers. This change represents the first
step in generating that callgraph - outputting a JSON structure to
Merlin's standard output file.

Next step will be to compare the outputs to something like JAM

Finally, note that I fixed a few subtle bugs when it came to running
merlin on a file by combining code paths and handling files that had
zero calls in them.

**Assumptions/Open issues:**

* We don't output a .json file at the moment for the callgraph, just a log of what went
  on in Merlin which is a text file that contains the full JSON output of the callgraph

* We can't yet run Merlin on the other benchmarks because it has no
  internal representation for built-in functions (like `console.log`)

**Testing:**

* Tested by running the experiments against various files with

**Issue**

N/A