Migrating eval-run.sh to Typescript/Javascript
**Motivation:**

The evaluation script is getting too long and complex to write in shell,
this changes it over to typescript/javascript which also allows us to do
more complex comparisons of the operations.

This also allows us to move the nodeprof-transform code into a more
general "evaluation" module, so we're not adding any more overhead than
was already there.

**Assumptions/Open issues:**

* This is a like-for-like move, subsequent PRs will build on the changes

* some of the debug scripts and the setup have been left as shell so
  that the migration is less impactful

**Testing:**

* running the evaluation and install scripts

**Issue**

* N/A